Fixes bug where data attributes are not properly initialized
The problem would occur when a slider was initialized using the data
attributes such as these

    data-slider-min="0" data-slider-max="32" data-slider-value="0"

The code uses the jQuery `.data` method to pull in these attributes. The
alternative would be to use the `.attr` method in which can also read
these data attributes. However, the `.attr` will only return a string
whereas we will get actual arrays and numbers back with `.data`.

Here is the previous code:

    this.min = this.element.data('slider-min') || options.min;
	this.max = this.element.data('slider-max') || options.max;
	this.step = this.element.data('slider-step') || options.step;
	this.value = this.element.data('slider-value') || options.value;

The above code attempts to pull the data attributes in if they are present
otherwise assigning the attributes to the default value sepecified in
`options`. The problem is that instead of getting a string back for
`slider-value` we get a number which evaluates to false in the
short circuting assignment used here, Setting our slider's initial
value to the default of `5` instead of the `0` we wanted.

The solution is more careful checking:

    this.min = this.element.data('slider-min');
	if (typeof this.min === 'undefined') this.min = options.min;

	this.max = this.element.data('slider-max');
	if (typeof this.max === 'undefined') this.max = options.max;

	this.step = this.element.data('slider-step');
	if (typeof this.step === 'undefined') this.step = options.step;

	this.value = this.element.data('slider-value');
	if (typeof this.value === 'undefined') this.value = options.value;